+++ lib/engineyard-serverside/deploy_hook.rb
@@ -3,21 +3,23 @@ require 'engineyard-serverside/shell/helpers'
 module EY
   module Serverside
     class DeployHook
+      attr_reader :config, :shell, :hook_name
+
       def initialize(config, shell, hook_name)
         @config, @shell, @hook_name = config, shell, hook_name
       end
 
       def hook_path
-        @config.paths.deploy_hook(@hook_name)
+        config.paths.deploy_hook(hook_name)
       end
 
       def callback_context
-        @context ||= CallbackContext.new(@config, @shell, hook_path)
+        @context ||= CallbackContext.new(config, shell, hook_path)
       end
 
       def call
         if hook_path.exist?
-          Dir.chdir(@config.paths.active_release.to_s) do
+          Dir.chdir(config.paths.active_release.to_s) do
             if desc = syntax_error(hook_path)
               hook_name = hook_path.basename
               abort "*** [Error] Invalid Ruby syntax in hook: #{hook_name} ***\n*** #{desc.chomp} ***"
@@ -29,14 +31,24 @@ module EY
       end
 
       def eval_hook(code)
+        display_deprecation_warnings(code)
         callback_context.instance_eval(code)
       rescue Exception => exception
         display_hook_error(exception, code, hook_path)
         raise exception
       end
 
+      def display_deprecation_warnings(code)
+        if code =~ /@configuration/
+          shell.warning("Use of `@configuration` in deploy hooks is deprecated.\nPlease use `config`, which provides access to the same object.\n\tin #{hook_path}")
+        end
+        if code =~ /@node/
+          shell.warning("Use of `@node` in deploy hooks is deprecated.\nPlease use `config.node`, which provides access to the same object.\n\tin #{hook_path}")
+        end
+      end
+
       def display_hook_error(exception, code, hook_path)
-        @shell.fatal <<-ERROR
+        shell.fatal <<-ERROR
 Exception raised in deploy hook #{hook_path}.
 
 #{exception.class}: #{exception.to_s}
@@ -53,13 +65,13 @@ Please fix this error before retrying.
       class CallbackContext
         include EY::Serverside::Shell::Helpers
 
-        attr_reader :shell
+        attr_reader :shell, :hook_path
 
         def initialize(config, shell, hook_path)
           @configuration = config
           @configuration.set_framework_envs
           @shell = shell
-          @node = node
+          @node = config.node
           @hook_path = hook_path
         end
 
@@ -73,6 +85,7 @@ Please fix this error before retrying.
 
         def method_missing(meth, *args, &blk)
           if @configuration.respond_to?(meth)
+            shell.warning "Use of `#{meth}` (via method_missing) is deprecated in favor of `config.#{meth}` for improved error messages and compatibility.\n\tin #{hook_path}"
             @configuration.send(meth, *args, &blk)
           else
             super